bot/pyproject.toml
[project]
name = "lms-telegram-bot"
version = "0.1.0"
description = "Telegram bot for interacting with the LMS backend"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["handlers"]

[tool.hatch.build.targets.sdist]
include = ["bot.py", "config.py", "handlers/", "pyproject.toml"]